Fix: Space.removeShape trashing arbiter list.
When a collision is ignored (returning false on the begin callback), the related arbiter doesn't make into the arbiter list of the body. But when the shape is removed, the space tries to unthread this ignored arbiter off the body arbiter list, trashing it.
I'm not sure if this is a good solution because I could'nt really grasp all the code. But this seems to work for me.